refactor: Renaming by removing enum
